test(health/coverage): parser tests, biomarker tests, integration test + ruff fixes
- 23 parser tests covering LCOV/Cobertura/Clover happy paths, edge
  cases (missing end_of_record, bad XML), format detection, and the
  test-file heuristic (paths + framework imports + paired-test lookup).
- 11 coverage-biomarker tests covering untested_hotspot fall-back when
  no coverage data, gating by dependents/hotspot, and coverage_gap
  severity / test-file exemption.
- Integration test runs the full HealthAnalyzer over the sample_repo
  fixture with a handcrafted LCOV report and asserts coverage flows
  through onto HealthFileMetric + coverage_gap biomarker fires.
- Ruff cleanups (RUF046, SIM105, SIM108, RUF003) on the new modules.

+++ packages/core/src/repowise/core/analysis/health/biomarkers/coverage_gap.py
@@ -39,7 +39,7 @@ class CoverageGapDetector:
         uncovered = total - len(ctx.covered_lines or ())
         # Some parsers don't emit a per-line set; fall back to math.
         if uncovered <= 0:
-            uncovered = int(round(total * (100.0 - cov) / 100.0))
+            uncovered = round(total * (100.0 - cov) / 100.0)
 
         deep_gap = cov < _LINE_COVERAGE_DEEP_GAP and total >= _MIN_FILE_SIZE
         regular_gap = cov < _LINE_COVERAGE_GAP and uncovered >= _MIN_UNCOVERED_LINES
